reorganize join
